% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{apply2dataset}
\alias{apply2dataset}
\title{apply2dataset}
\usage{
apply2dataset(executor, fun, workspace = NULL, dataset = NULL,
  element = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
apply2dataset
}
