% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignObj.R
\name{AssignObj}
\alias{AssignObj}
\title{Reassign an R object from one environment to another}
\usage{
AssignObj(obj, name = NULL, from = NULL, to = NULL, remove = T)
}
\arguments{
\item{obj}{an R symbol.}

\item{name}{optional new name of the R object whithin its destination environment.}

\item{from}{source environment where the object is currently located.}

\item{to}{destination environment.}

\item{remove}{logical indicating if the original object should be removed at its current
location (default = T, yes).}
}
\description{
Reassign an R object from one environment to another
}
\seealso{
\link{MakeObj},
  \link{LoadObj},
  \link{DeleteObj}
}
