% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvailableObj.R
\name{AvailableObj}
\alias{AvailableObj}
\title{Check if the RDS file associated to an R object is available}
\usage{
AvailableObj(obj, path = NULL, name = NULL, relative = NULL)
}
\arguments{
\item{obj}{an R object.}

\item{path}{folder location for RDS files. This can be interpreted either as
a sub-directory of the \code{rootpath} location, or as an independent
location (see \code{relative}). By default \code{path} is unspecified.}

\item{name}{\strong{RESERVED FOR INTERNAL USE}: optional name of the R object.
When omitted this name is automatically provided by the symbol being passed
as first argument.}

\item{relative}{logical value controlling if the \code{path} option is interpreted as a
relative location, meaning as a sub-directory of the \code{rootpath}
location (default = \code{TRUE}, yes), or as an independent one.}
}
\value{
\code{TRUE} when the RDS file exists, \code{FALSE} otherwise.
}
\description{
Check if the RDS file associated to an R object is available
}
\details{
When unspecified, the value of the following argument(s) are determined
by the corresponding automation option(s) (see \link{LittleThumb}):

\code{relative}
}
\examples{
# Create a simple R object named 'a'
a <- 1:10

# Save object a in the current working directory
SaveObj(a)

# Simulate restarting the R environment
rm(a)

# Check if object a is available in the R environment (FALSE)
exists("a")

# Check if an RDS file is available for object a (TRUE)
AvailableObj(a)

# Load object a from the current working directory
LoadObj(a)

print(a)

# Delete the RDS file associated to object a and remove the object itself
DeleteObj(a)

# Now both lines below should return FALSE
AvailableObj(a)
exists("a")

# A verbose way to automatically make/save/load the object 'a' (don't use this)
if(AvailableObj(a)) {
  LoadObj(a)
} else {
  a <- 1:10
  SaveObj(a)
}

# Automatically make/save/load the object 'a'
MakeObj(a, { a <- 1:10 })

# Cleanup
DeleteObj(a)
}
\seealso{
\link{MakeObj},
  \link{LittleThumb}
}
\keyword{internal}
