% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_R_Objects.R
\name{check_class}
\alias{check_class}
\title{check_class}
\usage{
check_class(x, class, lst = F, strict = T, error = F)
}
\arguments{
\item{x}{a list of R objects or an R object}

\item{class}{\code{character} name}

\item{lst}{\code{logical} if TRUE then x is tested otherwise elements of x}

\item{strict}{\code{logical} if TRUE then x must be member of all classes}
}
\value{
check_class returns a \code{logical} value
}
\description{
verify if R object belong to requested classes
}
