% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_Basics.R
\name{checklist}
\alias{checklist}
\title{checklist}
\usage{
checklist(chk, lst, msg = "", spc = "\\n")
}
\arguments{
\item{chk}{logical vector}

\item{lst}{character vector}

\item{msg}{message}
}
\description{
checklist
}
