% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{clone_dataset}
\alias{clone_dataset}
\title{clone_dataset}
\usage{
clone_dataset(workspace, dataset, name, path = NULL, files = NULL,
  file_columns = NULL)
}
\arguments{
\item{file_columns}{}
}
\description{
clone_dataset
}
