% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(workspace, source_path = NULL, files = NULL,
  pattern = NULL, annotations = NULL, id_column = "name",
  file_columns = "file", name = NULL, path = NULL, delete_source = F,
  ask = T)
}
\arguments{
\item{ask}{}
}
\description{
Create dataset definition and copy associated data inside a workspace folder
}
