% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LittleThumb.R
\name{DefaultArgs}
\alias{DefaultArgs}
\title{DefaultArgs ** RESERVED FOR INTERNAL USE **}
\usage{
DefaultArgs(f, cfg, ignore = NULL, env = NULL)
}
\description{
DefaultArgs ** RESERVED FOR INTERNAL USE **
}
\keyword{internal}
