% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LittleThumb.R
\name{DefaultOptions}
\alias{DefaultOptions}
\title{DefaultOptions ** RESERVED FOR INTERNAL USE **}
\usage{
DefaultOptions()
}
\description{
DefaultOptions ** RESERVED FOR INTERNAL USE **
}
\keyword{internal}
