% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Workspace.R
\name{define_workspace}
\alias{define_workspace}
\title{define_workspace}
\usage{
define_workspace(name, path)
}
\arguments{
\item{path}{}
}
\description{
Add {name, path, is_created, is_opened} to LTE$workspaces register
}
