% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteObj.R
\name{DeleteObj}
\alias{DeleteObj}
\title{Delete the RDS file associated to an R object}
\usage{
DeleteObj(obj, path = NULL, name = NULL, relative = NULL, envir = NULL,
  remove = NULL, messages = NULL, ...)
}
\arguments{
\item{obj}{symbol corresponding to an R object previously made by \link{MakeObj}
or saved using \link{SaveObj}.}

\item{path}{directory where the RDS file should be located.}

\item{name}{\strong{RESERVED FOR INTERNAL USE}: optional name of the R object.
When omitted this name is automatically provided by the symbol being passed
as first argument.}

\item{relative}{logical value controlling whether the specified \code{path} should be
absolute or relative to the global \code{path} option accessible via the
\link{LittleThumb} function (default = T, yes).}

\item{envir}{\link{environment} where the R object should be located.}

\item{remove}{logical.}

\item{messages}{logical value enabling or disabling status messages from LittleThumb
functions (default = T, yes).}

\item{...}{optional arguments passed to the \link{rm} function.}
}
\description{
Delete the RDS file associated to an R object
}
\examples{
# Create a simple R object named 'a'
a <- 1:10

# Save object a in the current working directory
SaveObj(a)

# Simulate restarting the R environment
rm(a)

# Check if object a is available in the R environment (FALSE)
exists("a")

# Check if an RDS file is available for object a (TRUE)
AvailableObj(a)

# Load object a from the current working directory
LoadObj(a)

print(a)

# Delete the RDS file associated to object a and remove the object itself
DeleteObj(a)

# Now both lines below should return FALSE
AvailableObj(a)
exists("a")

# A verbose way to automatically make/save/load the object 'a' (don't use this)
if(AvailableObj(a)) {
  LoadObj(a)
} else {
  a <- 1:10
  SaveObj(a)
}

# Automatically make/save/load the object 'a'
MakeObj(a, { a <- 1:10 })

# MkObj is a short alias of function MakeObj
MkObj(a, { a <- 1:10 })

# Cleanup
DeleteObj(a)
}
\seealso{
\link{LittleThumb},
  \link{MakeObj},
  \link{LoadObj},
  \link{SaveObj},
  \link{AvailableObj}
}
