% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_accessors.R
\name{GetValue}
\alias{GetValue}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
GetValue(x, prn.name = NULL, parent = NULL, origin = parent.frame())
}
\description{
** RESERVED FOR INTERNAL USE **
}
\keyword{internal}
