% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_Basics.R
\name{list_modules}
\alias{list_modules}
\title{list_modules}
\usage{
list_modules(...)
}
\arguments{
\item{...}{arguments passed to the \link{dir} function}
}
\value{
module names
}
\description{
list_modules
}
