% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(workspace = NULL, dataset = NULL, element = NULL,
  reader = read.delim, ...)
}
\arguments{
\item{...}{}
}
\description{
Load data from workspace/dataset folders into the workspace environment
}
