% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_RData.R
\name{loadObj}
\alias{loadObj}
\title{load an R object from RData with automated path and file name}
\usage{
loadObj(obj, path = NULL, name = NULL, need = F, pos = .GlobalEnv,
  overload = F, ...)
}
\arguments{
\item{obj}{an R object previously saved using \link{saveObj}.}

\item{path}{base folder for saving R objects. By default, path to this folder is defined
by the value of \code{.Cfg$rdata.path} if such configuration variable exists.
Otherwise it is the current working directory when none is specified.}

\item{name}{a file name (default = none).}

\item{need}{logical, when \code{TRUE} loadObj raises an error if object loading fails
(default = F).}

\item{pos}{the target \link{environment} where the object should be loaded
(default = .GlobalEnv).
See \link{assign} for documentation on the different ways to specify
environments with the \code{pos} paramter.}

\item{...}{optional parameters passed to the \link{readRDS} function.}
}
\value{
logical
\code{TRUE} when object was loaded successfully, \code{FALSE} otherwise.
}
\description{
load an R object from RData with automated path and file name
}
\examples{
# R objects are automatically saved in the current working directory
a <- rep("a", 10)
saveObj(a)
rm(a)
loadObj(a)
print(a)
file.remove("a.rdata") # delete the file generated for this example

# R objects can be saved and loaded from any path
b <- rep("b", 10)
saveObj(b, path = "MyRData") # path is automatically created if necessary
b <- NULL
loadObj(b, path = "MyRData")

# Not yet defined R objects may be created only when no saved version exists
if(! loadObj(a, "MyRData")) {
  a <- 0
  saveObj(a, "MyRData")
}
print(a)

# Saved object can be updated by overwriting
a <- 1
saveObj(a, path = "MyRData")
rm(a)
loadObj(a)
print(a)
}
\seealso{
\link{saveObj},
  \link{rdataPath},
  \link{readRDS},
  \link{reassign}
}
