% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_dependencies.R
\docType{data}
\name{LT}
\alias{LT}
\title{** RESERVED FOR INTERNAL USE **}
\format{An object of class \code{environment} of length 1.}
\usage{
LT
}
\description{
** RESERVED FOR INTERNAL USE **
}
\keyword{internal}
