% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT_Core.R
\docType{class}
\name{LT_Description-class}
\alias{LT_Description-class}
\alias{properties,LT_Description-method}
\alias{property,LT_Description,character-method}
\alias{datakeys,LT_Description-method}
\alias{ndata,LT_Description-method}
\title{LT_Description}
\usage{
\S4method{properties}{LT_Description}(obj)

\S4method{property}{LT_Description,character}(obj, key)

\S4method{datakeys}{LT_Description}(obj)

\S4method{ndata}{LT_Description}(obj)
}
\arguments{
\item{obj}{object}

\item{key}{property name}
}
\value{
\code{properties} returns a \code{character} vector

\code{datakeys} returns a \code{character} vector

\code{ndata} returns an \code{integer}
}
\description{
This S4 class implements an hybrid data structure combining
a list and a data frame.
}
\section{Methods (by generic)}{
\itemize{
\item \code{properties}: top level names of the properties list

\item \code{property}: accessor for top level properties

\item \code{datakeys}: column names of the description data frame

\item \code{ndata}: number of rows in the description data frame
}}

