% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_FileSystem.R
\name{lt_path}
\alias{lt_path}
\title{lt_path}
\usage{
lt_path(obj, ...)
}
\arguments{
\item{obj}{either an LT_Config, LT_Workspace or LT_Dataset object}
}
\value{
lt_path returns the path to the object's self definition
}
\description{
self defintion path of an object
}
