% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Channel.R, R/LTC_Workspace.R
\name{LT_Workspace}
\alias{LT_Workspace}
\alias{LT_Workspace}
\title{LT_Channel}
\usage{
LT_Workspace(...)

LT_Workspace(...)
}
\arguments{
\item{...}{}

\item{...}{}
}
\value{
an object of class LT_Channel

an object of class LT_Dataset
}
\description{



}
