% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_TextRepresentation.R
\name{make_id}
\alias{make_id}
\title{make_id}
\usage{
make_id(n = 1, tag = "LT")
}
\arguments{
\item{n}{number of identifiers to be generated}

\item{tag}{prefix of generated identifiers}
}
\value{
\code{character}
}
\description{
make_id
}
