% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_FileSystem.R
\name{make_path.character}
\alias{make_path.character}
\title{make_path}
\usage{
\method{make_path}{character}(..., ext = "")
}
\arguments{
\item{...}{components of a file path}

\item{ext}{file name extension (default = "")}
}
\value{
make_path returns a \code{character} value
}
\description{
make_path
}
