% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_FileSystem.R
\name{make_path.data.frame}
\alias{make_path.data.frame}
\title{make_path}
\usage{
\method{make_path}{data.frame}(obj, paths, name, root = "", key = 1,
  value = 2)
}
\arguments{
\item{obj}{data.frame containing two columns used to define paths as {key, value} pairs}

\item{paths}{list defining a file system tree structure}

\item{name}{key of the path to be build}

\item{root}{root location of the path to be build}

\item{key}{column name for path keys in the obj data.frame}

\item{value}{column name for path values in the obj data.frame}
}
\value{
make_path returns a character value representing the requested file system
path.
}
\description{
Path construction using depth search first.
Note: Does not support multiple occurence of the same path key in the path
structure.
}
