% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePath.R
\name{MakeDir}
\alias{MakeDir}
\title{Create a destination path (Linux/macOS only)}
\usage{
MakeDir(x)
}
\description{
\code{MakeDir} (short alias \link{MkDir}) creates a destination path
including upper level directories if necessary.
}
\keyword{internal}
