% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePath.R
\name{MakePath}
\alias{MakePath}
\title{Concatenate several strings to form a filesystem path}
\usage{
MakePath(..., ext = NULL, relative = NULL)
}
\arguments{
\item{...}{list of character strings forming a file path.}

\item{ext}{file name extension (default = "", none).}

\item{relative}{TODO: documentation}
}
\value{
\code{MakePath} returns a \code{character} string.
}
\description{
\link{MkPath} is a short alias of function \code{MakePath}.
}
\keyword{internal}
