% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{ManageObjectAndParentArgs}
\alias{ManageObjectAndParentArgs}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ManageObjectAndParentArgs(a, origin = parent.frame())
}
\description{
List of arguments involving a delayed expression evaluation.
}
\keyword{internal}
