% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_accessors.R
\name{ObjectExists}
\alias{ObjectExists}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ObjectExists(x, prn.name = NULL, parent = NULL, origin = parent.frame(),
  inherits = FALSE)
}
\description{
** RESERVED FOR INTERNAL USE **
}
\keyword{internal}
