% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{ObjWithExpressionArgs}
\alias{ObjWithExpressionArgs}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ObjWithExpressionArgs(a, xpr, explicit = "name")
}
\description{
List of arguments involving a delayed expression evaluation
}
\keyword{internal}
