% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTE.R
\name{openLittleThumb}
\alias{openLittleThumb}
\title{openLittleThumb}
\usage{
openLittleThumb()
}
\description{
Load or create the LittleThumb environement (LTE), which contains
configuration options as well as the workspace, dataset and job registers.
}
