% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_RData.R
\name{rdataPath}
\alias{rdataPath}
\title{default path for saving R objects}
\usage{
rdataPath(path = NULL, name = NULL)
}
\arguments{
\item{path}{base folder for saving R objects. By default, path to this folder is defined
by the value of \code{.Cfg$rdata.path} if such configuration variable exists.
Otherwise it is the current working directory when none is specified.}

\item{name}{a file name (default = none).}
}
\value{
path
}
\description{
default path for saving R objects
}
\seealso{
\link{saveObj},
  \link{loadObj}
}
