% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_R_Objects.R
\name{reassign}
\alias{reassign}
\title{Reassign object to a different environment}
\usage{
reassign(obj, pos, src = globalenv(), keep = F)
}
\arguments{
\item{obj}{an R object or the name of an R object}

\item{pos}{specifies a target \link{environment} where the object should be reassigned.
See \link{assign} for documentation on the different ways to specify
environments with the \code{pos} paramter.}

\item{src}{specifies the source \link{environment} where the object is currently located
(default = .GlobalEnv).
See \link{assign} for documentation on the different ways to specify
environments with the \code{pos} paramter.}

\item{keep}{logical indicating if the object should be removed or preserved
at its current location (defautl = F, remove).}
}
\description{
copy or move an R object from the source environment \code{src} to the target
environment \code{pos}.
}
\seealso{
\link{assign},
  \link{environment}
}
