% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_accessors.R
\name{RemoveObject}
\alias{RemoveObject}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
RemoveObject(x, prn.name = NULL, parent = NULL,
  origin = parent.frame())
}
\description{
** RESERVED FOR INTERNAL USE **
}
\keyword{internal}
