% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{save_data}
\alias{save_data}
\title{save_data}
\usage{
save_data(workspace = NULL, dataset = NULL, element = NULL,
  writer = write.table, ...)
}
\arguments{
\item{...}{}
}
\description{
Save data from the workspace environment into workspace/dataset folders
}
