% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveObj.R
\name{SaveObj}
\alias{SaveObj}
\title{Save an R object as RData with semi-automated path and file name}
\usage{
SaveObj(obj, path = NULL, name = NULL, env = NULL, ...)
}
\arguments{
\item{obj}{an R object to be saved as RData.}

\item{path}{specific location used to save the RData.
When none is specified this location is defined by the path option accessible
via \link{LittleThumb}, which by default is the current working directory.}

\item{name}{optional file name. When omitted this name is automatically set to the
name of the R object being passed as argument.}

\item{env}{the \link{environment} containing the R object to be saved
(default = .GlobalEnv).}

\item{...}{optional parameters passed to the \link{saveRDS} function.}
}
\description{
Save an R object as RData with semi-automated path and file name
}
\examples{
# R objects are automatically saved in the current working directory
a <- rep("a", 10)
SaveObj(a)
rm(a)
LoadObj(a)
print(a)
file.remove("a.rdata") # delete the file generated for this example

# R objects can be saved and loaded from any path
b <- rep("b", 10)
SaveObj(b, path = "MyRData") # path is automatically created if necessary
b <- NULL
LoadObj(b, path = "MyRData")

# Not yet defined R objects may be created only when no saved version exists
if(! LoadObj(a, "MyRData")) {
  a <- 0
  SaveObj(a, "MyRData")
}
print(a)

# Saved object can be updated by overwriting
a <- 1
SaveObj(a, path = "MyRData")
rm(a)
LoadObj(a)
print(a)
}
\seealso{
\link{MakeObj},
  \link{LoadObj},
  \link{AvailableObj}
}
