% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTF_FileSystem.R
\name{self_path}
\alias{self_path}
\title{self_path}
\usage{
self_path(obj, ...)
}
\arguments{
\item{obj}{either an LT_Workspace or LT_Dataset object}
}
\value{
self_path returns the path to the workspace or dataset
}
\description{
path to workspaces and datasets
}
