% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LittleThumb.R
\name{StatusMessage}
\alias{StatusMessage}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
StatusMessage(status, name = NULL, file = NULL, sep = NULL)
}
\description{
** RESERVED FOR INTERNAL USE **
}
\keyword{internal}
