% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{which_dataset}
\alias{which_dataset}
\title{which_dataset}
\usage{
which_dataset(id = NULL, workspace = NULL, name = NULL)
}
\arguments{
\item{name}{}
}
\value{
integer
}
\description{
Return the register index of datasets
}
