% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Workspace.R
\name{which_workspace}
\alias{which_workspace}
\title{which_dataset}
\usage{
which_workspace(name = NULL)
}
\arguments{
\item{name}{}
}
\value{
integer
}
\description{
Return the register index of workspaces
}
