% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Workspace.R
\name{workspace_object}
\alias{workspace_object}
\title{workspace_object}
\usage{
workspace_object(name = NULL)
}
\arguments{
\item{name}{}
}
\value{
object of class LT_Workspace
}
\description{
Return a registered workspace object
}
