% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignObj.R
\name{AssignObj}
\alias{AssignObj}
\title{Reassign an R object from its parent environment to another one}
\usage{
AssignObj(obj, name = NULL, from = NULL, to = NULL, remove = TRUE)
}
\arguments{
\item{obj}{an R object.}

\item{name}{\strong{RESERVED FOR INTERNAL USE}: optional name of the R object.
When omitted this name is automatically provided by the symbol being passed
as first argument using \link{substitute} and \link{deparse}.}

\item{from}{source environment, where the object is currently located.}

\item{to}{destination environment.}

\item{remove}{logical value controlling wheter the object should removed from the source
environment (default = \code{TRUE}, yes).}
}
\description{
Reassign an R object from its parent environment to another one
}
\seealso{
\link{MakeObj},
  \link{LittleThumb}
}
