% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvailableObj.R
\name{AvailableObj}
\alias{AvailableObj}
\title{Check if an R object is available as RData}
\usage{
AvailableObj(obj, path = NULL, name = NULL)
}
\arguments{
\item{obj}{an R object previously generated with \link{MakeObj} or saved using
\link{SaveObj}.}

\item{path}{specific location where to look for the RData.
When none is specified this location is defined by the path option accessible
via \link{LittleThumb}, which by default is the current working directory.}

\item{name}{optional file name. When omitted this name is automatically set to the
name of the R object being passed as argument.}
}
\value{
logical, \code{TRUE} when RData is available and \code{FALSE} otherwise.
}
\description{
Check if an R object is available as RData
}
\seealso{
\link{LoadObj},
  \link{MakeObj},
  \link{SaveObj}
}
\keyword{internal}
