% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{clone_dataset}
\alias{clone_dataset}
\title{clone_dataset}
\usage{
clone_dataset(workspace, dataset, name, path = NULL, files = NULL,
  file_columns = NULL)
}
\arguments{
\item{workspace}{name}

\item{dataset}{name}

\item{name}{new dataset}

\item{path}{new path}

\item{files}{new files}

\item{file_columns}{new file column name}
}
\description{
clone_dataset
}
