% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(workspace, source_path = NULL, files = NULL,
  pattern = NULL, annotations = NULL, id_column = "name",
  file_columns = "file", name = NULL, path = NULL, delete_source = F,
  ask = T)
}
\arguments{
\item{workspace}{name}

\item{source_path}{path}

\item{files}{character vector}

\item{pattern}{regular expression}

\item{annotations}{path}

\item{id_column}{column name}

\item{file_columns}{column name}

\item{name}{dataset name}

\item{path}{base path of dataset files}

\item{delete_source}{logical (default = F)}

\item{ask}{logical (default = T)}
}
\description{
Create dataset definition and copy associated data inside a workspace folder
}
