% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{dataset_object}
\alias{dataset_object}
\title{dataset_object}
\usage{
dataset_object(id = NULL, workspace = NULL, name = NULL)
}
\arguments{
\item{name}{}
}
\value{
object of class LT_Dataset
}
\description{
Return registered dataset names
}
