% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{DefaultArgs}
\alias{DefaultArgs}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
DefaultArgs(cfg, ignore = NULL, fun = NULL, env = NULL)
}
\description{
Provide default values to unspecified arguments
}
\keyword{internal}
