% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTC_Dataset.R
\name{list_datasets}
\alias{list_datasets}
\title{list_datasets}
\usage{
list_datasets(workspace = NULL, detailed = F, x = NULL)
}
\arguments{
\item{workspace}{name}

\item{detailed}{logical (default = F)}

\item{x}{not implemented}
}
\value{
dataset names
}
\description{
Return registered dataset names
}
