% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LittleThumb.R
\name{LittleThumb}
\alias{LittleThumb}
\title{Automation options}
\usage{
LittleThumb(...)
}
\arguments{
\item{...}{Any of the following arguments.}

\item{rootpath}{root location for RDS files. By default \code{rootpath} is the current
working directory (see \link{getwd}).}

\item{path}{folder location for RDS files. This can be interpreted either as
a sub-directory of the \code{rootpath} location, or as an independent
location (see \code{relative}). By default \code{path} is unspecified.}

\item{extension}{\strong{DO NOT MODIFY}: RDS file extension (default = ".rds").}

\item{relative}{logical value controlling if the \code{path} option is interpreted as a
relative location, meaning as a sub-directory of the \code{rootpath}
location (default = \code{TRUE}, yes), or as an independent one.}

\item{parent}{\link{environment} where the R object should be located.
By default \code{parent} is the \link{parent.frame} of the function
being called.}

\item{makedir}{logical value controlling if non-existing folders should be created
automatically when necessary (default = \code{TRUE}, yes).}

\item{reload}{logical value forcing to load R objects from associated RDS files
even when these objects already exist in the R environment
(default = \code{FALSE}, no).}

\item{rebuild}{logical value forcing \code{MakeObj} to regenerate R objects even when
the associated RDS files already exist (default = \code{FALSE}, no).}

\item{cleanup}{logical value controlling wheter \code{MakeObj} should remove additional
objects encapsulated in the provided expression after its evaluation
(default = \code{TRUE}, yes).}

\item{remove}{logical value controlling wheter \code{DeleteObj} should remove the R object
after deleting the RDS file (default = \code{TRUE}, yes).}

\item{messages}{logical value enabling or disabling status messages from LittleThumb
functions (default = \code{TRUE}, enabled).}
}
\description{
This function sets the default value of arguments used by the main functions
of the LittleThumb package.
}
