% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LittleThumb.R
\name{LittleThumb}
\alias{LittleThumb}
\title{Global options for LittleThumb functions}
\usage{
LittleThumb(...)
}
\arguments{
\item{...}{Any of the following arguments:}

\item{path}{default folder used to save R objects.}

\item{extension}{RDS file extension (default = ".rds").}

\item{relative}{logical value. When \code{TRUE} the default folder defines the root path for
all read/write operations automated by LittleThumb, meaning that any path
specified when calling \code{MakeObj}, \code{SaveObj} or \code{LoadObj}
will represent a sub path of the default folder.}

\item{envir}{environment containing the R object.
With the default value (\code{NA}), this environment is the
\link{parent.frame} of called LittleThumb functions.}

\item{makedir}{logical value, if \code{TRUE} non-existing folders are created automatically
when necessary.}

\item{reload}{logical value, if \code{TRUE} forces to load R objects from associated RDS
files even when these objects already exist in the R environment
(default = F, no).}

\item{rebuild}{logical value, if \code{TRUE} forces \code{MakeObj} to regenerate R objects
even when the associated RDS files already exist (default = F, no).}

\item{cleanup}{logical value controlling wheter \code{MakeObj} should remove additional
objects created while evaluating the provided expression (default = T, yes).}

\item{remove}{logical value controlling wheter \code{DeleteObj} should remove the R object
after deleting the RDS file (default = T, yes).}

\item{messages}{logical value enabling or disabling status messages from LittleThumb
functions (default = T, enabled).}

\item{history}{TODO: implementation & documentation}
}
\description{
This function sets the default value of arguments used by the main functions
of the LittleThumb package, which allows to control the automated
saving/loading operations performed by these functions at the global level.
}
\seealso{
\link{MakeObj},
  \link{LoadObj},
  \link{SaveObj},
  \link{DeleteObj}
}
