% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeObj.R
\name{MakeObj}
\alias{MakeObj}
\title{Automatically make & save or load an R object}
\usage{
MakeObj(...)
}
\arguments{
\item{...}{at least an R symbol and an expression defining the corresponding R object.
The symbol must be provided in first position and the expression in last
position, both as unnamed arguments.}
}
\description{
\code{MakeObj} automatically generates an R object using the provided
expression and saves the associated RDS file, or skips these two operations
and loads the previously saved RDS file when it is available, or even avoids
to do anything when the object is already defined in its parent environment.

The automation performed by \code{MakeObj} can be controlled by optional
arguments and their corresponding global options which are accessible
via the \link{LittleThumb} function (see the details section).
}
\details{
The \code{MakeObj} function requires at least two arguments: an R symbol
in the first position and an R expression in the last position,
both being provided as unnamed arguments.

When evaluated, the provided expression must define the value
of the R object to be automatically saved or loaded.
For instance, the following line would automatically make/save/load
an integer vector named \code{x} and containing values from 1 to 10:

\code{MakeObj(x, { x <- 1:10 })}

In addition, the following named arguments can be used in between
the symbol and expression arguments (see \link{LittleThumb}):

\code{path}, \code{relative}, \code{embedded},  \code{parent},
\code{makedir}, \code{reload}, \code{rebuild}, \code{cleanup},
and \code{messages}
}
\note{
\link{MkObj} can be used as a short alias of function \code{MakeObj}.
}
\examples{
# Create a simple R object named 'a'
a <- 1:10

# Save object a in the current working directory
SaveObj(a)

# Simulate restarting the R environment
rm(a)

# Check if object a is available in the R environment (FALSE)
exists("a")

# Check if an RDS file is available for object a (TRUE)
AvailableObj(a)

# Load object a from the current working directory
LoadObj(a)

print(a)

# Delete the RDS file associated to object a and remove the object itself
DeleteObj(a)

# Now both lines below should return FALSE
AvailableObj(a)
exists("a")

# A verbose way to automatically make/save/load the object 'a' (don't use this)
if(AvailableObj(a)) {
  LoadObj(a)
} else {
  a <- 1:10
  SaveObj(a)
}

# Automatically make/save/load the object 'a'
MakeObj(a, { a <- 1:10 })

# Cleanup
DeleteObj(a)
}
\seealso{
\link{LittleThumb},
  \link{DeleteObj}
}
\keyword{visible}
