% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeObj.R
\name{MakeObj}
\alias{MakeObj}
\title{MakeObj}
\usage{
MakeObj(obj, ...)
}
\arguments{
\item{obj}{an R object.}

\item{...}{at least an expression defining the R object, which can be preceded by
optional parameters passed to the \link{LoadObj} and \link{SaveObj}
functions (see detail and example sections).}
}
\description{
Generate an R object and save it as RData, or skip the generation and load
this object from previously saved RData if available.
}
\details{
TODO: documentation
}
\seealso{
\link{LoadObj},
  \link{SaveObj},
  \link{AvailableObj}
}
