% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePath.R
\name{MakePath}
\alias{MakePath}
\title{Concatenate several strings to form a filesystem path}
\usage{
MakePath(..., ext = "")
}
\arguments{
\item{...}{list of character strings forming a file path.}

\item{ext}{file name extension (default = "", none).}
}
\value{
\code{MakePath} returns a \code{character} string.
}
\description{
Concatenate several strings to form a filesystem path
}
\keyword{internal}
